requirements.txt
# https://pypi.org/project/flask
Flask==3.0.2
setuptools
# https://pypi.org/project/gunicorn
gunicorn==21.2.0


google-generativeai==0.5.1
google-cloud-storage==2.9.0
Pillow==11.1.0.0
google-cloud-secret-manager==2.16.2